matchers: fix mmk_that not working
The broken code tried to stuff the predicate after the matcher
structure, which could have worked if we didn't actually copy the
matcher struct internally, dropping the predicate pointer.  What we
ended up with was pure garbage.

Now, we do it the "proper" way, by introducing a data pointer in the
matcher structure.