mock,api: use comparison operators rather than mmk_memcmp
mmk_memcmp might have been a good idea in theory.

That is, if you don't account for endianness.

We now use regular comparison operators.  This means that users can't use
comparison matchers with types that doesn't implement those operators,
but in practice this is exactly what mmk_that is for, so we choose
correctness over genericity.