win32: fixed vfprintf not being found for vitals.
Since the Universal CRT was introduced, vfprintf may no longer be
included in the runtime dll, but may be defined as an inline function
that calls __stdio_common_vfprintf instead.

Since the vitals.c code was searching for the vfprintf function, aborting
if not found, it was crashing under windows versions shipping UCRT. We now
search for __stdio_common_vfprintf as a fallback.