Flush CPU caches after mprotect().
If caches are not flushed, executing trampoline code on aarch64
often results in a SIGILL. This appears to be related to
execution permissions on trampoline pages and the CPU instruction
cache. On a related note, glibc's _dl_relocate_object() API does
clear caches when bringing code sections into memory space.

Signed-off-by: Michel Hidalgo <michel@ekumenlabs.com>